Move to float64

+++ src/github.com/alberthier/pathtracer/camera.go
@@ -2,11 +2,11 @@ package pathtracer
 
 type Camera struct {
 	Position Vector3
-	AngleX   float32
-	AngleY   float32
-	AngleZ   float32
+	AngleX   float64
+	AngleY   float64
+	AngleZ   float64
 }
 
-func NewCamera(x float32, y float32, z float32, angleX float32, angleY float32, angleZ float32) *Camera {
+func NewCamera(x float64, y float64, z float64, angleX float64, angleY float64, angleZ float64) *Camera {
 	return &Camera{Vector3{x, y, z}, angleX, angleY, angleZ}
 }
